[FLINK-29377][rpc] Reusable timeout extraction
